+++ tomcat/src/main/java/com/techcourse/servlet/DefaultServlet.java
@@ -1,8 +1,8 @@
 package com.techcourse.servlet;
 
 import org.apache.catalina.servlet.HttpServlet;
-import org.apache.coyote.http11.request.HttpServletRequest;
-import org.apache.coyote.http11.response.HttpServletResponse;
+import org.apache.coyote.http.request.HttpServletRequest;
+import org.apache.coyote.http.response.HttpServletResponse;
 import org.apache.util.FileUtils;
 
 public class DefaultServlet extends HttpServlet {
