+++ tomcat/src/main/java/com/techcourse/servlet/GreetingServlet.java
@@ -1,8 +1,8 @@
 package com.techcourse.servlet;
 
 import org.apache.catalina.servlet.HttpServlet;
-import org.apache.coyote.http11.request.HttpServletRequest;
-import org.apache.coyote.http11.response.HttpServletResponse;
+import org.apache.coyote.http.request.HttpServletRequest;
+import org.apache.coyote.http.response.HttpServletResponse;
 
 public class GreetingServlet extends HttpServlet {
