+++ tomcat/src/main/java/org/apache/coyote/http/HttpProtocol.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.common;
+package org.apache.coyote.http;
 
 import com.techcourse.exception.UncheckedServletException;
 import java.util.Arrays;
