+++ tomcat/src/main/java/org/apache/coyote/http/request/line/RequestLine.java
@@ -1,6 +1,6 @@
-package org.apache.coyote.http11.request.line;
+package org.apache.coyote.http.request.line;
 
-import org.apache.coyote.http11.common.HttpProtocol;
+import org.apache.coyote.http.HttpProtocol;
 
 public class RequestLine {
