+++ tomcat/src/main/java/org/apache/coyote/http/response/line/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.response.line;
+package org.apache.coyote.http.response.line;
 
 public enum HttpStatus {
