+++ tomcat/src/test/java/com/techcourse/servlet/mapping/RequestMappingTest.java
@@ -1,18 +1,18 @@
 package com.techcourse.servlet.mapping;
 
-import static org.apache.coyote.http11.common.HttpProtocol.HTTP_11;
-import static org.apache.coyote.http11.request.line.Method.GET;
+import static org.apache.coyote.http.HttpProtocol.HTTP_11;
+import static org.apache.coyote.http.request.line.Method.GET;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.techcourse.servlet.DefaultServlet;
 import com.techcourse.servlet.LoginServlet;
-import org.apache.coyote.http11.common.HttpHeaders;
-import org.apache.coyote.http11.common.HttpMessageBody;
-import org.apache.coyote.http11.common.HttpProtocol;
-import org.apache.coyote.http11.request.HttpServletRequest;
-import org.apache.coyote.http11.request.line.Method;
-import org.apache.coyote.http11.request.line.RequestLine;
-import org.apache.coyote.http11.request.line.Uri;
+import org.apache.coyote.http.HttpHeaders;
+import org.apache.coyote.http.HttpMessageBody;
+import org.apache.coyote.http.HttpProtocol;
+import org.apache.coyote.http.request.HttpServletRequest;
+import org.apache.coyote.http.request.line.Method;
+import org.apache.coyote.http.request.line.RequestLine;
+import org.apache.coyote.http.request.line.Uri;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
